Safely drop object contexts in the event of multiple threads
All WDF object context spaces are supposed to be thread-safe, which
means that we're not necessarily guaranteed exclusive access by the time
we need to unload the driver. This is mainly a problem for when we need
to unload a driver, as properly managing object refcounts means that by
the time `EvtDestroyCallback` is called, there are supposed to be no
other object handles present.

The initialization flag also serves another purpose as a way of preventing
accidentally dropping context spaces that weren't fully initialized, but
this can only happen in the context of unwinding panics.