Define `WdfMinimumVersionRequired` in `lib.rs`
Originally it was defined in `wrapper.c`, but after some investigation spawned
off from someone on RPLCS (thanks!), it turns out that `wrapper.c` wasn't
needed anymore.

The original purpose of `wrapper.c` was to make the WDF inline function stubs
usable in Rust. This also lead to a duplicate definition of
`WdfMinimumVersionRequired` between the `lib.rs` and in the `wrapper_bindings`
lib, so the one in `lib.rs` was removed in 9aa4f45a36d4cfc2c9a5d017f2fbc976429fa07e

Due to *vibes*, manually dispatching to the WDF functions themselves was
preferred instead, and `wrapper.c` wasn't needed anymore after that.

This also removes `windows-kernel-sys`'s `wrapper.c`, which was originally
intended to hold wrappers around `ProbeForRead` and `ProbeForWrite`.
Fortunately WDF provides equivalent functionality in the form of
`WdfRequestProbeAndLockUserBufferForRead` and
`WdfRequestProbeAndLockUserBufferForWrite`, so it's not needed anymore too.