Decouple the `Sync` and `Send`ness of handles and context spaces
Handles are (in theory) `Send + Sync` and context spaces are required to be
`Sync` (in theory). Handles also give out references to the context spaces
and does not have direct ownership of them, so we use `fn() -> T` for the
`PhantomData` type instead of `T`.

Also makes `Mutex` `Send` since it doesn't have the requirement of the
storage not moving.