Use `get_context_space_ptr` in `context_pin_init`
Was previously manually getting the context space and then doing a cast to `T`,
which is technically incorrect as we store the drop flag too. This didn't
matter too much though as the drop flag was stored after the data so that
either way, the pointer ends up in the right place.