Start new `ndis-rs` library
Initially starting out with porting things that we want from
`ndis-driver-library`, but intended to eventually encompass things like nbl
pool creation and (very later on) serve as the foundation for implementing
common protocol/miniport/intermediate driver boilerplate.

Theoretically it's testable as-is, but due to the hard-coded linker flags in
`driver/.cargo/config.toml`, it's trying to compile the test binary as if
it were trying to compile the driver binary.