Add `Driver::get`
The only place that the driver handle is uninitialized is before calling
`Driver::create`, where it obviously wouldn't make sense to call it.

Unfortunately, we still had one instance where the driver context could be
accessed while it was still initializing (the EvtIo* callbacks), so we moved
the control device initialization to later. It's still not known whether this
is ever a possible situation, but better to leave a note for later.